+++ app/csv/csv_column_stats.py
@@ -1,4 +1,4 @@
-from app.utils.csv.type_parser import TypeParser, TYPE_PARSER_HIERARCHY
+from app.csv.type_parser import TypeParser, TYPE_PARSER_HIERARCHY
 
 
 class CsvColumnStats:
