+++ app/csv/csv_utils.py
@@ -1,7 +1,7 @@
 import csv
 import sys
 from typing import IO
-from app.utils.csv import CsvColumnStats
+from app.csv import CsvColumnStats
 
 
 def derive_csv_column_types(file: IO, max_row_count: int = sys.maxsize):
