+++ app/tests/csv/tests_type_parser.py
@@ -1,7 +1,7 @@
 import datetime
 from unittest import TestCase
 
-from app.utils.csv import parse_boolean, parse_date, parse_datetime, parse_time
+from app.csv import parse_boolean, parse_date, parse_datetime, parse_time
 
 
 class TypeParserTests(TestCase):
